copy-n-paste mania for tracks and overlays
git-svn-id: http://svn.code.sf.net/p/qlandkartegt/code/QLandkarteGT/trunk@2215 d0c010d8-bb82-477e-a075-112140230451

+++ src/COverlayDistance.h
@@ -109,7 +109,7 @@ class COverlayDistance : public IOverlay
         /// the distance in [m]
         double distance;
         /// optional speed in km/s, <=0 is no speed
-        double speed;
+        float speed;
 
         /// rectangle in function wheel for del icon
         QRect rectDel;
